--- spinifex/migrate/predastoretopology/testdata/predastore_v4_multinode.toml
@@ -1,97 +1,0 @@
-# Predastore S3-compatible storage configuration file in TOML format.
-# Generated for multi-node cluster
-version = "4"
-region = "ap-southeast-2"
-
-host = "0.0.0.0"
-port = 8443
-
-debug = false
-
-[rs]
-data = 2
-parity = 1
-
-[[db]]
-id = 1
-host = "10.0.0.1"
-port = 6660
-path = "distributed/db/node-1/"
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-leader = true
-
-[[db]]
-id = 2
-host = "10.0.0.2"
-port = 6660
-path = "distributed/db/node-2/"
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-
-[[db]]
-id = 3
-host = "10.0.0.3"
-port = 6660
-path = "distributed/db/node-3/"
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-
-[[nodes]]
-id = 1
-host = "10.0.0.1"
-port = 9991
-path = "distributed/nodes/node-1/"
-
-[[nodes]]
-id = 2
-host = "10.0.0.2"
-port = 9991
-path = "distributed/nodes/node-2/"
-
-[[nodes]]
-id = 3
-host = "10.0.0.3"
-port = 9991
-path = "distributed/nodes/node-3/"
-
-[[buckets]]
-name = "predastore"
-region = "ap-southeast-2"
-type = "distributed"
-public = false
-encryption = ""
-account_id = "000000000000"
-
-# Northstar DNS zone storage. Read-only for the northstar service; the control
-# plane writes zone files with the system credentials.
-[[buckets]]
-name = "northstar-zones"
-region = "ap-southeast-2"
-type = "distributed"
-public = false
-encryption = ""
-account_id = "000000000000"
-
-[[auth]]
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-account_id = "000000000000"
-policy = [
-  { bucket = "predastore", actions = ["s3:ListBucket", "s3:GetObject", "s3:PutObject", "s3:DeleteObject", "s3:ListAllMyBuckets"] },
-  { bucket = "northstar-zones", actions = ["s3:ListBucket", "s3:GetObject", "s3:PutObject", "s3:DeleteObject"] },
-]
-
-# API request throttling (per-account, per-action token bucket)
-[ratelimit]
-enabled = false
-rate = 100          # Default sustained requests per second per account per action
-burst = 500         # Default maximum burst capacity
-
-# IAM authentication via NATS KV (enables multi-account S3 access)
-[iam]
-nats_url = "nats://10.0.0.2:4222"
-nats_token = "s3cr3t"
-nats_ca_cert = "/etc/spinifex/ca.pem"
-master_key_path = "/etc/spinifex/master.key"
-access_keys_bucket = "spinifex-iam-access-keys"
