--- spinifex/migrate/predastoretopology/testdata/predastore_v4_singlenode.toml
@@ -1,109 +1,0 @@
-# Predastore S3-compatible storage configuration file in TOML format.
-# MULTI NODE used for testing
-version = "4"
-region = "ap-southeast-2"
-
-# Optionally configure port and host, or specify via daemon CLI options
-host = "0.0.0.0"
-port = 8443
-
-# Enable verbose debugging
-debug = false
-
-# Optionally configure base-path if full directory path not defined for "fs" buckets
-# base_path = "/home/predastore/data/"
-
-# Specify reed solomon encoding -
-# RS(2,1), 2 data, 1 parity. Minimum
-# RS(3,2), 3 data, 2 parity. Recommended for production
-[rs]
-data = 2
-parity = 1
-
-[[db]]
-id = 1
-host = "127.0.0.1"
-port = 6660
-#raft_port = 7660       # Optional: Raft consensus port (default: port + 1000)
-#advertise_host = ""    # Optional: Address to advertise to other Raft nodes
-path = "distributed/db/node-1/"
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-leader = true      # Bootstrap leader hint
-
-[[db]]
-id = 2
-host = "127.0.0.1"
-port = 6661
-path = "distributed/db/node-2/"
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-
-[[db]]
-id = 3
-host = "127.0.0.1"
-port = 6662
-path = "distributed/db/node-3/"
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-
-# Shard nodes
-# Using distributed hash ring for object > shard placement.
-[[nodes]]
-id = 1
-host = "127.0.0.1"
-port = 9991
-path = "distributed/nodes/node-1/"
-
-[[nodes]]
-id = 2
-host = "127.0.0.1"
-port = 9992
-path = "distributed/nodes/node-2/"
-
-[[nodes]]
-id = 3
-host = "127.0.0.1"
-port = 9993
-path = "distributed/nodes/node-3/"
-
-[[buckets]]
-name = "predastore"
-region = "ap-southeast-2"
-type = "distributed"
-public = false
-encryption = ""
-account_id = "000000000000"
-
-[[auth]]
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-account_id = "000000000000"
-policy = [
-  { bucket = "predastore", actions = ["s3:ListBucket",  "s3:GetObject", "s3:PutObject", "s3:DeleteObject", "s3:ListAllMyBuckets"] },
-]
-
-# API request throttling (per-account, per-action token bucket)
-[ratelimit]
-enabled = false
-rate = 100          # Default sustained requests per second per account per action
-burst = 500         # Default maximum burst capacity
-
-[ratelimit.action."s3:PutObject"]
-rate = 50
-burst = 200
-
-[ratelimit.action."s3:DeleteObject"]
-rate = 50
-burst = 200
-
-# IAM authentication via NATS KV (enables multi-account S3 access)
-[iam]
-nats_url = "nats://localhost:4222"
-nats_token = "s3cr3t"
-nats_ca_cert = "/etc/spinifex/ca.pem"
-master_key_path = "/etc/spinifex/master.key"
-access_keys_bucket = "spinifex-iam-access-keys"
-
-[compaction]
-interval_seconds = 300
