--- spinifex/migrate/predastoretopology/testdata/predastore_v5_multinode.golden
@@ -1,109 +1,0 @@
-# Predastore S3-compatible storage configuration file in TOML format.
-# Generated for multi-node cluster
-version = "5"
-region = "ap-southeast-2"
-
-# The gateway listen address comes from the daemon CLI options: a top-level
-# host key would collide with the [[host]] topology tables below.
-
-debug = false
-
-[rs]
-data = 2
-parity = 1
-
-# One host per machine: a predastore process owning a socket and a data
-# directory. bind_addr binds every interface; public_addr is what the other
-# hosts dial.
-
-[[host]]
-id = 1
-bind_addr = "0.0.0.0:6660"
-public_addr = "10.0.0.1:6660"
-data_dir = "{{DATA_DIR}}/predastore/cluster"
-
-[[host]]
-id = 2
-bind_addr = "0.0.0.0:6660"
-public_addr = "10.0.0.2:6660"
-data_dir = "{{DATA_DIR}}/predastore/cluster"
-
-[[host]]
-id = 3
-bind_addr = "0.0.0.0:6660"
-public_addr = "10.0.0.3:6660"
-data_dir = "{{DATA_DIR}}/predastore/cluster"
-
-# Nodes are roles pinned to a host: shard-storage nodes hold erasure-coded
-# object shards, state replicas form the Raft quorum over global state.
-
-[[node]]
-id = 1
-host_id = 1
-role = "shard-storage"
-
-[[node]]
-id = 2
-host_id = 2
-role = "shard-storage"
-
-[[node]]
-id = 3
-host_id = 3
-role = "shard-storage"
-
-[[node]]
-id = 4
-host_id = 1
-role = "state-replica"
-
-[[node]]
-id = 5
-host_id = 2
-role = "state-replica"
-
-[[node]]
-id = 6
-host_id = 3
-role = "state-replica"
-
-[[buckets]]
-name = "predastore"
-region = "ap-southeast-2"
-type = "distributed"
-public = false
-encryption = ""
-account_id = "000000000000"
-
-# Northstar DNS zone storage. Read-only for the northstar service; the control
-# plane writes zone files with the system credentials.
-[[buckets]]
-name = "northstar-zones"
-region = "ap-southeast-2"
-type = "distributed"
-public = false
-encryption = ""
-account_id = "000000000000"
-
-[[auth]]
-access_key_id = "AKIAIOSFODNN7EXAMPLE"
-secret_access_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
-account_id = "000000000000"
-policy = [
-  { bucket = "predastore", actions = ["s3:ListBucket", "s3:GetObject", "s3:PutObject", "s3:DeleteObject", "s3:ListAllMyBuckets"] },
-  { bucket = "northstar-zones", actions = ["s3:ListBucket", "s3:GetObject", "s3:PutObject", "s3:DeleteObject"] },
-]
-
-# API request throttling (per-account, per-action token bucket)
-[ratelimit]
-enabled = false
-rate = 100          # Default sustained requests per second per account per action
-burst = 500         # Default maximum burst capacity
-
-# IAM authentication via NATS KV (enables multi-account S3 access)
-[iam]
-nats_url = "nats://10.0.0.2:4222"
-nats_token = "s3cr3t"
-nats_ca_cert = "/etc/spinifex/ca.pem"
-master_key_path = "/etc/spinifex/master.key"
-access_keys_bucket = "spinifex-iam-access-keys"
